+++ NEWS.md
@@ -1,3 +1,8 @@
+# EpiInvert 0.1.2
+
+* We include the management of weekly aggregated incidence data where every 
+week a single data is communicated with the accumulated incidence in the last 7 days. 
+
 # EpiInvert 0.1.1
 
 * README.md has been simplified to avoid problems with the render of mathematical